requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
Django==3.2.9
django-bootstrap4==21.1
django-cors-headers==3.10.0
django-datetimepicker==3.14
django-extra-views==0.14.0
djangorestframework==3.12.4
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
yapf==0.31.0
